@@ -36,7 +36,7 @@ file(MAKE_DIRECTORY ${INSTALL_BIN_PATH})
 
 set(QT_LOCATOR_CMAKE "${CMAKE_CURRENT_SOURCE_DIR}/QtLocator.cmake")
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 link_directories(${CMAKE_LIBRARY_OUTPUT_DIRECTORY})
