@@ -92,18 +92,33 @@ bool Person::load(const QJsonObject& obj)
 }
 bool Person::save(QJsonObject& obj) const
 {
-    obj["firstName"] = firstName.c_str();
-    obj["lastName"] = lastName.c_str();
-    obj["gender"] = gender.c_str();
-    obj["age"] = age.c_str();
-    obj["email"] = email.c_str();
-    obj["phone"] = phone.c_str();
-    obj["education"] = education.c_str();
-    obj["occupation"] = occupation.c_str();
-    obj["experience"] = experience.c_str();
-    obj["salary"] = salary.c_str();
-    obj["numberOfChildren"] = numberOfChildren.c_str();
-    obj["martialStatus"] = martialStatus.c_str();
+
+   // obj["firstName"] = firstName.c_str();
+   // obj["lastName"] = lastName.c_str();
+   // obj["gender"] = gender.c_str();
+   // obj["age"] = age.c_str();
+   // obj["email"] = email.c_str();
+   // obj["phone"] = phone.c_str();
+   // obj["education"] = education.c_str();
+   // obj["occupation"] = occupation.c_str();
+   // obj["experience"] = experience.c_str();
+   // obj["salary"] = salary.c_str();
+   // obj["numberOfChildren"] = numberOfChildren.c_str();
+   // obj["martialStatus"] = martialStatus.c_str();
+
+
+    obj.insert("firstName", firstName.c_str());
+    obj.insert("lastName", lastName.c_str());
+    obj.insert("gender", gender.c_str());
+    obj.insert("age", age.c_str());
+    obj.insert("email", email.c_str());
+    obj.insert("phone", phone.c_str());
+    obj.insert("education", education.c_str());
+    obj.insert("occupation", occupation.c_str());
+    obj.insert("experience", experience.c_str());
+    obj.insert("salary", salary.c_str());
+    obj.insert("numberOfChildren", numberOfChildren.c_str());
+    obj.insert("martialStatus", martialStatus.c_str());
     return true;
 }
 
