@@ -2,10 +2,12 @@
 
 #include "JD_base.h"
 #include "JDDeclaration.h"
+#include "manager/async/WorkProgress.h"
 #include <vector>
 #include <mutex>
 #include <condition_variable>
 #include <atomic>
+#include <memory>
 
 namespace JsonDatabase
 {
@@ -19,33 +21,37 @@ namespace JsonDatabase
                 std::mutex& mtx);
             ~JDManagerAsyncWorker();
             void setup();
-            void addWork(JDManagerAysncWork* work);
-            bool isWorkDone(JDManagerAysncWork* work);
-            void removeDoneWork(JDManagerAysncWork* work);
+            void addWork(std::shared_ptr<JDManagerAysncWork> work);
+            bool isWorkDone(std::shared_ptr<JDManagerAysncWork> work);
+            void removeDoneWork(std::shared_ptr<JDManagerAysncWork> work);
             void clearDoneWork();
             void process();
             void start();
             void stop();
         public:
 
-
-            
+            WorkProgress getWorkProgress() const;
+            bool isBusy() const;
+            WorkType getCurrentWorkType() const;
 
         private:
             void threadLoop();
-            void processWork(const std::vector<JDManagerAysncWork*>& workList);
-            void processWork(JDManagerAysncWork* work);
+            void processWork(const std::vector<std::shared_ptr<JDManagerAysncWork>>& workList);
+            void processWork(std::shared_ptr<JDManagerAysncWork> work);
 
             JDManager& m_manager;
             std::mutex& m_mutex;
 
             std::thread* m_thread;
-            std::vector<JDManagerAysncWork*> m_workList;
-            std::vector<JDManagerAysncWork*> m_workListDone;
+            std::vector<std::shared_ptr<JDManagerAysncWork>> m_workList;
+            std::vector<std::shared_ptr<JDManagerAysncWork>> m_workListDone;
             std::mutex m_mutexInternal;
+            std::mutex m_workListMutex;
             std::condition_variable m_cv;
             std::atomic<bool> m_stopFlag;
+            std::atomic<bool> m_busy;
 
+            std::atomic<std::shared_ptr<JDManagerAysncWork>> m_currentWork;
             
         };
     }
